+++ scripts/session_restorer.sh
@@ -109,10 +109,10 @@ restore_all_sessions() {
 restore_all_pane_processes() {
 	if restore_pane_processes_enabled; then
 		local pane_full_command
-		awk 'BEGIN { FS="\t"; OFS="\t" } /^pane/ && $11 !~ "^:$" { print $2, $3, $7, $11; }' $(last_session_path) |
-			while IFS=$'\t' read session_name window_number pane_index pane_full_command; do
+		awk 'BEGIN { FS="\t"; OFS="\t" } /^pane/ && $11 !~ "^:$" { print $2, $3, $7, $8, $11; }' $(last_session_path) |
+			while IFS=$'\t' read session_name window_number pane_index dir pane_full_command; do
 				pane_full_command="$(remove_first_char "$pane_full_command")"
-				restore_pane_process "$pane_full_command" "$session_name" "$window_number" "$pane_index"
+				restore_pane_process "$pane_full_command" "$session_name" "$window_number" "$pane_index" "$dir"
 			done
 	fi
 }